Two player identification
Added functionality to track player sessions. The server will
automatically generate a single map for both players when it has
detected that two players have joined. You can test it by using two
separate browsers (for instance FF and Chrome). You have to manually
refresh the browser in order to get the map - it will not autoupdate.

Additionally: added javadocs for javax,servlet to make it easier to
write code. I could not make it be relative so you might have to
manually fix the reference as seen here:
https://www.dropbox.com/s/2vufquq2576f6iy/1.png